// src/gen/ship.ts
export interface BoundingBox {
    /**
     * @generated from protobuf field: double minLat = 1
     */
    minLat: number;
    /**
     * @generated from protobuf field: double minLon = 2
     */
    minLon: number;
    /**
     * @generated from protobuf field: double maxLat = 3
     */
    maxLat: number;
    /**
     * @generated from protobuf field: double maxLon = 4
     */
    maxLon: number;
}
/**
 * @generated from protobuf message ship.SubscribeRequest
 */
export interface SubscribeRequest {
    /**
     * @generated from protobuf field: ship.BoundingBox box = 1
     */
    box?: BoundingBox;
}
// ...
class BoundingBox$Type extends MessageType<BoundingBox> {
    constructor() {
        super("ship.BoundingBox", [
            { no: 1, name: "minLat", kind: "scalar", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 2, name: "minLon", kind: "scalar", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 3, name: "maxLat", kind: "scalar", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 4, name: "maxLon", kind: "scalar", T: 1 /*ScalarType.DOUBLE*/ }
        ]);
    }
    create(value?: PartialMessage<BoundingBox>): BoundingBox {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.minLat = 0;
        message.minLon = 0;
        message.maxLat = 0;
        message.maxLon = 0;
        if (value !== undefined)
            reflectionMergePartial<BoundingBox>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: BoundingBox): BoundingBox {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* double minLat */ 1:
                    message.minLat = reader.double();
                    break;
                case /* double minLon */ 2:
                    message.minLon = reader.double();
                    break;
                case /* double maxLat */ 3:
                    message.maxLat = reader.double();
                    break;
                case /* double maxLon */ 4:
                    message.maxLon = reader.double();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: BoundingBox, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* double minLat = 1; */
        if (message.minLat !== 0)
            writer.tag(1, WireType.Bit64).double(message.minLat);
        /* double minLon = 2; */
        if (message.minLon !== 0)
            writer.tag(2, WireType.Bit64).double(message.minLon);
        /* double maxLat = 3; */
        if (message.maxLat !== 0)
            writer.tag(3, WireType.Bit64).double(message.maxLat);
        /* double maxLon = 4; */
        if (message.maxLon !== 0)
            writer.tag(4, WireType.Bit64).double(message.maxLon);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message ship.BoundingBox
 */
export const BoundingBox = new BoundingBox$Type();
// @generated message type with reflection information, may provide speed optimized methods
class SubscribeRequest$Type extends MessageType<SubscribeRequest> {
    constructor() {
        super("ship.SubscribeRequest", [
            { no: 1, name: "box", kind: "message", T: () => BoundingBox }
        ]);
    }
    create(value?: PartialMessage<SubscribeRequest>): SubscribeRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        if (value !== undefined)
            reflectionMergePartial<SubscribeRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SubscribeRequest): SubscribeRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* ship.BoundingBox box */ 1:
                    message.box = BoundingBox.internalBinaryRead(reader, reader.uint32(), options, message.box);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SubscribeRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* ship.BoundingBox box = 1; */
        if (message.box)
            BoundingBox.internalBinaryWrite(message.box, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message ship.SubscribeRequest
 */
export const SubscribeRequest = new SubscribeRequest$Type();